[refactor] : TimeUtil 테스트 가능하게 수정 및 Proxy 삭제
